Moves stacked area graph to its own file.
Not so sure about this move. Splitting these graphs into their
own files makes the files shorter, but it sort of makes sense
to put the graph holder and all the graphs in one file where
they can all have access to the same imported values.

Thoughts anyone?
